Replaced "Reset" with "Skip"
I consider skipping the current iteration a much more relevant action,
than resetting the timer. If someone, for example, has a scheduled break
or wants to take a break after finishing a demanding task, it is
necessary to skip the current task.

In the rare occasions that require for resetting the timer, one could
do so manually after pausing, or simply skip twice.