Layout of the TimerView GUI has been improved
The clock is neatly centered and the buttons are arranged in a way, that
I consider visually and semantically appealing. As Done is the most
"destructive" action, I placed this button in the left corner.
Start/Pause and Reset are located on the right, because they somewhat
belong together as they both control the "flow" of the timing mechanism.