Added default timer durations to SettingsManager
I improved the get_value function of SettingsManager. It now accepts an
optional "default" argument. Thus it is possible to define the "get",
"set" and "default" behavior of a settings attribute in the same place.

The new settings, "task_duration" and "break_duration", can be
adjusted in the "Timer" section of the configuration file.